extracting source/resources required for the GradleRunner tests
