+++ src/test/groovy/org/gradle/android/writer/JavaTest.groovy
@@ -1,0 +1,13 @@
+package org.gradle.android.writer
+
+class JavaTest {
+
+    static String simpleTest(String packageName){
+        return """
+                package ${packageName};
+
+                public class JavaUserTest {
+                }
+            """.stripIndent()
+    }
+}
