+++ src/test/groovy/org/gradle/android/writer/Kotlin.groovy
@@ -1,0 +1,13 @@
+package org.gradle.android.writer
+
+class Kotlin {
+
+    static String simpleDataClass(String packageName) {
+        return """
+                package ${packageName}
+
+                data class Foo(val lable: String)
+
+            """.stripIndent()
+    }
+}
